
\docType{data}
\name{isoTab}
\alias{isoTab}
\title{Stable isotope data}
\format{
A list containing
\describe{
  \item{name}{The element name: periodic table name (e.g. C, N, O, Mg, Ca, etc.)}
  \item{n}{number of stable isotopes}
  \item{masses}{an array containing the atomic masses of each of the different stable isotopes}
  \item{denomIso}{the reference isotope used to calculate stable isotope ratios}
  \item{IS}{The name of the international standard reference material}
  ...
}
}
\usage{
isoTab
}
\description{
Stable isotope data for the most common isotopic tracers in environmental studies, including: C, N, H, O, Mg, Ca, K, Li and B.
}
\keyword{datasets}
